requirements.txt
deep_translator==1.7.0
python-dotenv==0.19.2
requests==2.26.0
tweepy==4.12.1
